+++ src/common/vector_math.h
