+++ src/video_core/color.h
@@ -6,8 +6,7 @@
 
 #include "common/common_types.h"
 #include "common/swap.h"
-
-#include "video_core/math.h"
+#include "common/vector_math.h"
 
 namespace Color {
