+++ src/video_core/debug_utils/debug_utils.cpp
@@ -19,9 +19,9 @@
 #include "common/assert.h"
 #include "common/file_util.h"
 #include "common/math_util.h"
+#include "common/vector_math.h"
 
 #include "video_core/color.h"
-#include "video_core/math.h"
 #include "video_core/pica.h"
 #include "video_core/utils.h"
 #include "video_core/video_core.h"
